// tools/toml-to-yaml-converter/client/types.ts
export type TomlToYamlConverterMessages = Readonly<{
  meta: {
    name: string
    description: string
  }
  importFromFileLabel: string
  downloadYamlLabel: string
  tomlLabel: string
  tomlDescription: string
  yamlLabel: string
  yamlDescription: string
  yamlEmptyDescription: string
  tomlPlaceholder: string
  invalidTomlLabel: string
  copyYamlLabel: string
  copiedLabel: string
}>